requirements.txt
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface_hub
transformers
python-dotenv
numpy
scikit-learn
sentence-transformers
langchain-groq
ragas 
datasets 
pandas
tqdm
faiss-cpu
openpyxl
langchain-google-genai
evaluate
rouge_score
nltk
